Master->dev

+++ vagrant/vagrant-scripts/vagrant-cleanup.sh
@@ -1,11 +1,24 @@
 #!/bin/bash
 
-rm ../config/init.sh
+
+if [ -f ../config/init.sh ] ; then
+    rm ../config/init.sh
+fi
 touch ../config/init.sh
 
-rm ../config/cert.sh
+if [ -f ../config/cert.sh ] ; then
+    rm ../config/cert.sh
+fi
 touch ../config/cert.sh
 
 vagrant destroy -f
+
 rm -rf ../.vagrant
-rm .vagrant-state
+
+if [ -f .vagrant-state ] ; then
+    rm .vagrant-state
+fi
+
+if [ -f ../.vagrant-state ] ; then
+    rm ../.vagrant-state
+fi
